# docs/site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flare named-token API",
    "version": "1.0.0",
    "summary": "Account-scoped file uploads, file metadata, short links, and outbound file-ready events.",
    "description": "Documents the current unversioned routes in Flare, not an /api/v1 prefix. Replace the server URL with your own instance. Only named-token-supported endpoints appear under paths; browser-session/admin routes are documented separately. Bearer tokens use application scopes, not OAuth. Upload sizes are bytes; list-file sizes are MiB. Some responses intentionally omit success or data wrappers. The info version versions this document; webhook payload version is independently 1. Every operation requires both its token scope and the owner’s current role permission (x-flare-permissions), rechecked on each request. Role/user administration remains browser-session-only.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/FlintSH/Flare/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://files.example.com",
      "description": "Replace with your Flare instance origin."
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/files": {
      "get": {
        "operationId": "listFiles",
        "summary": "List and search your files",
        "tags": ["Files"],
        "description": "Requires the files:read scope. Pagination is offset-based except for image-anchor requests. File sizes in this response are MiB, not bytes. Results include only this account and use Cache-Control: private, no-store. The token owner must currently have files.read; role revocation applies on the next request.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-flare-scope": "files:read",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "cm_example",
                      "name": "screenshot.png",
                      "urlPath": "/abc123/screenshot.png",
                      "mimeType": "image/png",
                      "size": 0.1953125,
                      "uploadedAt": "2026-09-20T12:00:00.000Z",
                      "visibility": "PUBLIC",
                      "views": 0,
                      "downloads": 0,
                      "folderId": null,
                      "user": {
                        "urlId": "abc123"
                      },
                      "tags": [],
                      "hasPassword": false,
                      "expiresAt": null
                    }
                  ],
                  "pagination": {
                    "total": 1,
                    "pageCount": 1,
                    "page": 1,
                    "limit": 24
                  }
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "private, no-store"
              }
            }
          },
          "401": {
            "description": "Unauthorized: missing/invalid token, expired/revoked token, missing scope, or required email verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request/options, part list, file type, or expiration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Owned resource or upload session not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "The operation could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated account lacks the required current role permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "You do not have permission to perform this action."
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Positive page number.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Positive integer. Values above 100 are capped at 100.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 24
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive filename or OCR-text match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort order; unknown values fall back to newest.",
            "schema": {
              "type": "string",
              "default": "newest",
              "enum": [
                "newest",
                "oldest",
                "largest",
                "smallest",
                "name",
                "most-viewed",
                "least-viewed",
                "most-downloaded",
                "least-downloaded"
              ]
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "description": "Comma-separated exact MIME types, for example image/png,image/jpeg.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "description": "Inclusive uploaded-at lower bound; use an ISO 8601 timestamp with timezone.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "description": "Inclusive upper bound. A date-only value includes the final day in server-local time; explicit timestamps preserve their instant.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "description": "Comma-separated public, private, hasPassword. Conditions within this filter are ORed.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "description": "Owned folder ID or unfiled. Direct folder membership, without descendants.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "description": "Owned tag ID or untagged. Excluded tags do not count.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "galleryAnchor",
            "in": "query",
            "required": false,
            "description": "Image file ID in the current filter; requires galleryDirection. Returns adjacent images, excluding anchor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "galleryDirection",
            "in": "query",
            "required": false,
            "description": "Requires galleryAnchor. Anchored responses include pagination.offset and only image files.",
            "schema": {
              "type": "string",
              "enum": ["next", "previous"]
            }
          }
        ],
        "x-flare-permissions": ["files.read"]
      },
      "post": {
        "operationId": "uploadFile",
        "summary": "Upload one file",
        "tags": ["Files"],
        "description": "Requires the files:upload scope. Stream one multipart file. Select profile/folder in headers or query and naming in profile/account defaults before sending bytes. All upload limits and quota apply; accounts with quotas.bypass or Administrator are exempt from the default quota but not file-size limits. No multipart idempotency key is implemented. The token owner must currently have files.upload; role revocation applies on the next request. Without files.share, new uploads are private. Expiration requires files.delete for DELETE or files.share for SET_PRIVATE, including inherited defaults; missing action permission returns 403. Uploads applying tags/folders need tags.manage/folders.manage.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-flare-scope": "files:upload",
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UploadLinks"
                    },
                    "success": {
                      "const": true,
                      "type": "boolean"
                    }
                  },
                  "required": ["data", "success"]
                },
                "example": {
                  "success": true,
                  "data": {
                    "url": "https://files.example.com/abc123/screenshot.png",
                    "pageUrl": "https://files.example.com/abc123/screenshot.png",
                    "rawUrl": "https://files.example.com/api/files/abc123/screenshot.png",
                    "downloadUrl": "https://files.example.com/api/files/cm_example/download",
                    "copyText": "https://files.example.com/abc123/screenshot.png",
                    "name": "screenshot.png",
                    "size": 204800,
                    "type": "image/png"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: missing/invalid token, expired/revoked token, missing scope, or required email verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request/options, part list, file type, or expiration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Required current role permission is missing. Profile binding, ownership, or final token validity check failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Owned resource or upload session not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "File size, quota, or part-size limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The operation could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Upload-start rate limit: 30 requests per minute per client IP per application process.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Too many requests"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait; currently 60.",
                "schema": {
                  "type": "string"
                },
                "example": "60"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Upload-Profile",
            "in": "header",
            "required": false,
            "description": "Owned profile ID; none bypasses default. Must agree with query/body selection.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Upload-Folder",
            "in": "header",
            "required": false,
            "description": "Owned destination folder ID; none means unfiled. Must agree with query/body selection.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "required": false,
            "description": "Owned profile ID or none. Must agree with header.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folderId",
            "in": "query",
            "required": false,
            "description": "Owned folder ID or none. Must agree with header.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MultipartUploadRequest"
              }
            }
          }
        },
        "x-flare-permissions": ["files.upload"]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "flr_…",
        "description": "Named API token from Profile → Integrations. Select files:read, files:upload, urls:read, urls:write as needed. Required scope is recorded in each operation description and x-flare-scope; HTTP bearer security arrays are empty because this is not OAuth. Missing scopes produce 401. Required owner permission is recorded in x-flare-permissions. Missing current role permission produces 403 after authentication succeeds. Administrator never expands token scopes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"],
        "description": "Upload/authentication failures use this shape; no success property is required."
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "success": {
            "type": "boolean",
            "const": false
          }
        },
        "required": ["error", "success"]
      },
      "UploadLinks": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "pageUrl": {
            "type": "string",
            "format": "uri"
          },
          "rawUrl": {
            "type": "string",
            "format": "uri"
          },
          "downloadUrl": {
            "type": "string",
            "format": "uri"
          },
          "copyText": {
            "type": "string",
            "format": "uri"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "Bytes."
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "pageUrl",
          "rawUrl",
          "downloadUrl",
          "copyText",
          "name",
          "size",
          "type"
        ],
        "description": "URLs follow sharing access rules; they are not private-file access grants. url and copyText alias pageUrl. name is the display name; type is the MIME type.",
        "example": {
          "url": "https://files.example.com/abc123/screenshot.png",
          "pageUrl": "https://files.example.com/abc123/screenshot.png",
          "rawUrl": "https://files.example.com/api/files/abc123/screenshot.png",
          "downloadUrl": "https://files.example.com/api/files/cm_example/download",
          "copyText": "https://files.example.com/abc123/screenshot.png",
          "name": "screenshot.png",
          "size": 204800,
          "type": "image/png"
        }
      },
...
      "MultipartUploadRequest": {
        "type": "object",
        "properties": {
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "PRIVATE"],
            "description": "Public without a profile override. Private content requires an eligible browser session; this bearer token does not authorize content downloads."
          },
          "expiration": {
            "type": "string",
            "enum": ["DISABLED", "HOUR", "DAY", "WEEK", "MONTH"],
            "description": "Relative expiration; omission inherits. DISABLED explicitly disables it. MONTH increments the UTC calendar month."
          },
          "expiryAction": {
            "type": "string",
            "enum": ["DELETE", "SET_PRIVATE"],
            "description": "Action when expiration is processed. Inherits the profile/account default."
          },
          "randomizeFileUrls": {
            "type": "string",
            "enum": ["true", "false"],
            "description": "May only repeat the already selected naming option; set naming in the profile/account before upload."
          },
          "shareStyle": {
            "type": "string",
            "enum": ["minimal", "framed", "delivery"]
          },
          "tagIds": {
            "type": "string",
            "description": "JSON-encoded array of at most 20 owned tag IDs, for example [\"TAG_ID\"]."
          },
          "profileId": {
            "type": "string",
            "description": "Owned profile ID, or null to bypass the account default. A profile-bound token cannot switch profiles. Multipart uses an empty string or the literal string null for null. Select with X-Upload-Profile or the profileId query before sending bytes."
          },
          "folderId": {
            "type": "string",
            "description": "Owned destination folder ID, or null for unfiled. A request-only selection, fixed at initialization. Multipart uses an empty string or the literal string null for null. Select with X-Upload-Folder or the folderId query before sending bytes."
          },
          "password": {
            "type": "string",
            "description": "At most 72 UTF-8 bytes as well as 72 characters. Empty or null clears password protection. Not a saved profile option. Multipart uses an empty string or the literal string null for null."
          },
          "expiresAt": {
            "type": "string",
            "description": "Future ISO 8601 timestamp with timezone; null disables expiration. Overrides relative expiration. Forbidden for a profile-bound token. Multipart uses an empty string or the literal string null for null."
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "One file. Let the HTTP client create the multipart boundary."
          }
        },
        "required": ["file"],
        "description": "One multipart file with optional string fields. Profile/folder/naming must be selected before streaming starts."
      },
...
      "FileMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "urlPath": {
            "type": "string",
            "description": "Relative URL path based on the account URL ID."
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "minimum": 0,
            "description": "Size in MiB (bytes divided by 1,048,576). UploadLinks.size uses bytes."
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "PRIVATE"]
          },
          "views": {
            "type": "integer",
            "minimum": 0
          },
          "downloads": {
            "type": "integer",
            "minimum": 0
          },
          "folderId": {
            "type": ["string", "null"]
          },
          "user": {
            "type": "object",
            "properties": {
              "urlId": {
                "type": "string"
              }
            },
            "required": ["urlId"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": ["id", "name"]
            }
          },
          "hasPassword": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "urlPath",
          "mimeType",
          "size",
          "uploadedAt",
          "visibility",
          "views",
          "downloads",
          "folderId",
          "user",
          "tags",
          "hasPassword",
          "expiresAt"
        ],
        "description": "Owner-scoped metadata, including private files. Password hashes, content, and OCR text are not returned.",
        "example": {
          "id": "cm_example",
          "name": "screenshot.png",
          "urlPath": "/abc123/screenshot.png",
          "mimeType": "image/png",
          "size": 0.1953125,
          "uploadedAt": "2026-09-20T12:00:00.000Z",
          "visibility": "PUBLIC",
          "views": 0,
          "downloads": 0,
          "folderId": null,
          "user": {
            "urlId": "abc123"
          },
          "tags": [],
          "hasPassword": false,
          "expiresAt": null
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "pageCount": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "description": "Present for gallery-anchor navigation."
          }
        },
        "required": ["total", "pageCount", "page", "limit"]
      },
      "FileListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileMetadata"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": ["success", "data", "pagination"]
      },
...
    }
  }
}